Introducing search/output management - render_search_result.py

+++ app.py
@@ -1,38 +1,33 @@
+# Native modules
 import os
-
+# External modules
 from flask import Flask, request, jsonify, render_template
 from py.seq_search import run as seq_search
-import random
-import string
-import datetime
+from py.seq_search import random_name_gen
 app = Flask(__name__)
-
-
-def random_name_gen():
-    # Generate Unique filename to store blastouts
-    n = 20
-    prefix = ''.join(random.choices(string.ascii_letters +
-                                    string.digits, k=n))
-    timestamp = datetime.datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
-    return f"{prefix}_{timestamp}"
 
 
 @app.route('/', methods=["POST", "GET"])
 def gfg():
     if request.method == "POST":
 
-        # getting input from search box
-        sequence = request.form["search-box"]
+        # Get input from search box
+        fasta_sequence = request.form["search-box"]
         seq_path = f"jobs/seqIN_{random_name_gen()}.fasta"
         with open(seq_path, 'w') as f:
-            f.write(sequence)
-        #
+            f.write(fasta_sequence)
+
+        # Test sequence format and perform sequence search if applicable
         blast_result = seq_search(seq_path)
-        os.remove(seq_path)
-        return render_template(blast_result)
+
+        # Deliver search output result
+        if blast_result:
+            os.remove(seq_path)
+            return render_template(blast_result)
+
     else:
         return render_template("index.html")
 
 
-if __name__ =='__main__':
+if __name__ == '__main__':
     app.run()
